AGREGADOS LISTA PRECIO Y PRECIO LISTA PRODUCTO

+++ src/main/java/com/techxperts/erp/producto/dto/ProductoDTO.java
@@ -18,7 +18,7 @@ public class ProductoDTO {
 
     private BigDecimal costo;
     private BigDecimal precio;
-    private BigDecimal descuentoMaximo;
+    private Integer descuentoMaximo;
     private BigDecimal precioOferta;
     private LocalDate fechaInicioOferta;
     private LocalDate fechaFinOferta;
